+++ Changes.md
@@ -19,6 +19,7 @@ Fixes
 API
 ---
 
+- Gaffer : Added ShufflePlug and ShufflesPlug which facilitate renaming key/value pairs (#3611).
 - Metadata : Added support for `renameable` metadata in BoxUI, LabelPlugValueWidget and NameWidget.
 - NodeEditor : Deprecated `setReadOnly()` and `getReadOnly()` methods.
 - NodeUI : Deprecated `setReadOnly()` and `getReadOnly()` methods.